Add files via upload
In this quad-rotor system we analysed Controllability and estimation of the system and compare with the results for real world. We find optimal ways to control the quad-rotor system for balancing at different thrusts and torques during the course of its motion. Also, we analysed linear system model of this quad-rotor at its equilibrium point. Moreover, we devised the non-linear system model for this system to analyze its controllibility and stability. We implemented Linear Quadratic Regulation and Kalman filter model to this linear system.

Also, there was no co-variance matrix of noise has been taken to update the kalman filter measurement equation.